requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
SQLAlchemy==2.0.23
bcrypt==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
Werkzeug==2.3.7
selenium==4.15.2
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
setuptools==80.9.0